requirements.txt
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0
Streamlit>= 1.9.2
scikit-learn==1.4.1.post1
tensorflow-cpu==2.16.1
keras==3.1.1
protobuf==4.25.3
altair==5.3.0
